SOACL_Web_Services/Flask_Web_Service/requirements.txt
Flask
Flask-restful
requests
flask_cors
mysql-connector